requirements.txt
streamlit==1.49.0
openai==1.102.0
python-dotenv==1.0.1
langdetect==1.0.9
speechrecognition==3.10.0
gtts==2.5.4
pillow==10.4.0
pyaudio==0.2.14
listen==0.1.6